Tuning tolerance of verability & Add of Functions
# Addition of
## shooterCheck
Checks for loose connections going to sparkMaxControllers
## manipulatorCheck
Checks for loose connections going to sparkMaxControllers
## lifterCheck
Checks for loose connections going to sparkMaxControllers

# Repairs to
## driveCheck
This was to correct for improper calculation of drive sparkMaxControllers being in tolerance for voltage/connection.
Checks current draw commaprison and checks if they are in tolerance of eachother on PowerHub.